Refactor deduplication (#159)
