+++ app/processing/transform-payment-request.js
@@ -6,32 +6,25 @@ const { confirmDueDates } = require('./due-dates')
 const { enrichPaymentRequests } = require('./enrichment')
 const { applyDualAccounting } = require('./dual-accounting')
 
-const transformPaymentRequest = async (paymentRequest, paymentRequests) => {
+const transformPaymentRequest = async (paymentRequest) => {
   // If BPS, then need to confirm payment request number as rekeyed claims can result in duplicate payment request numbers
   if (paymentRequest.schemeId === BPS) {
     paymentRequest.paymentRequestNumber = await confirmPaymentRequestNumber(paymentRequest)
   }
-
   // Check to see if payment request has had a previous payment request.
   // if yes, need to treat as post payment adjustment and calculate Delta which can result in payment request splitting
   const previousPaymentRequests = await getCompletedPaymentRequests(paymentRequest)
+  const sanitizedPaymentRequest = applyDualAccounting(paymentRequest, previousPaymentRequests)
 
-  // if no initial paymentRequests, apply dual accounting and set both deltaPaymentRequests and paymentRequests, or return if no previousPaymentRequests
-  let deltaPaymentRequests = null
-  if (!paymentRequests) {
-    const sanitizedPaymentRequest = applyDualAccounting(paymentRequest, previousPaymentRequests)
-    if (!previousPaymentRequests.length) {
-      return { completedPaymentRequests: [sanitizedPaymentRequest] }
-    }
-    deltaPaymentRequests = calculateDelta(sanitizedPaymentRequest, previousPaymentRequests)
-    paymentRequests = deltaPaymentRequests.completedPaymentRequests
+  if (previousPaymentRequests.length) {
+    const deltaPaymentRequests = calculateDelta(sanitizedPaymentRequest, previousPaymentRequests)
+    const confirmedPaymentRequests = confirmDueDates(deltaPaymentRequests.completedPaymentRequests, previousPaymentRequests)
+    deltaPaymentRequests.completedPaymentRequests = enrichPaymentRequests(confirmedPaymentRequests, previousPaymentRequests)
+
+    return deltaPaymentRequests
   }
 
-  const confirmedPaymentRequests = confirmDueDates(paymentRequests, previousPaymentRequests)
-  paymentRequests = enrichPaymentRequests(confirmedPaymentRequests, previousPaymentRequests)
-
-  // return the whole deltaPaymentRequests if exists or paymentRequests if not
-  return deltaPaymentRequests ?? paymentRequests
+  return { completedPaymentRequests: [sanitizedPaymentRequest] }
 }
 
 module.exports = {
